py_requirement.txt
requests
beautifulsoup4
tldextract
google-generativeai